Fixed to create output files at the relative path.
For example, if the input directory is /a and the md file was at
/a/b/c/x.md, then if the output dir is specified as /output, then the
output file will be located at /output/b/c/x.md, rather than
/output/x.md.

+++ src/metadata.cc
@@ -52,7 +52,7 @@ std::vector<std::string> SplitString(std::string_view line) {
 }  // namespace
 
 std::optional<Metadata> MetadataFactory::ParseMetadata(std::string_view content,
-                                                       int& end) {
+                                                       size_t& end) {
   size_t current = 0;
 
   // Ignore until it sees non whitespace.
